95-legal-brief-writer/requirements.txt
# Core
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI
click>=8.1.0
rich>=13.0.0

# Web UI
streamlit>=1.30.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
